Remove TopsyTurvy plugin.
On trying it out I can see that there's a basic flaw in my use.
I really don't want the switch to happen for number keys in all my
shortcuts.  If I want the "number key" shortcuts in command mode or the
like but the shifted version shortcuts for inserting text then I have to
make that distinction in software.

There's no way that the keyboard can tell what sort of thing I'm sending
keys to.