Go back to using the TopsyTurvy plugin
I've noted that it's utility is much less than I'd hoped (and it's not
going to replace the equivalent vim and emacs plugins any time soon).

However it does seem like it might be useful in limited circumstances,
and there's no reason to not have it on the keyboard in order to try and
look for those circumstances.