Hack to make layers work as I'd like them to
I want a "toggle shifted keyswitches" key.  I want that key to be on a
separate layer.

The current implementation I have for my shifted keyswitches is done via
layers, and the turn on/off behaviour is implemented via
activating/deactivating a special layer.

Rather than modify the layers.cpp plugin, I could instead make these
keyswitches check the state of a variable and have the "toggle shifted
keyswitches" button toggle that variable.

That would be less intrusive into code that is more fundamental across
the entire firmware but a little more ugly in the main file.

I actually think that would be a better solution, but this is the one I
thought of first, it's less work to type up, and it gives me an
opportunity to play around with changing the source-proper rather than
just the main file.