Don't initialise every element in macroStorage
I initialise the elements in macroStorage so that attempting to play
a macro which has not been recorded immediately hits a MACRO_ACTION_END
and is hence harmless.

Initialising *every* element in the buffer is not necessary for this.
Doing that seems like it might hide problems failing to properly
terminate macros until macros have been recorded and re-recorded.

Only initialising the first entry for each macro seems like it would
save a bit of compute power (though I don't think that's very
important), and also make it more likely I'll identify problems with
termination early on with testing.  Hence that's what I'm doing.