Move checking MACROREC into recordKeystroke
We need to record the MACRO_ACTION_END in our macro so we need to pass
the MACROREC key to our recordKeystroke function.  There is a small
edge-case to think about in handling a MACROREC while having just
requested that we wait for the "next macro to play".  That is now
handled in recordKeystroke since the implementation is to record
something.