Attempt to move MacrosOnTheFly into a separate plugin
This should not only make it easier to use between the Model100 and
Model01 setups, but it should also make it easier to test.

The testing is the thing that I'm most interested in at the moment.  I'd
like to get started testing this so that I can start to understand the
behaviour I'm expecting to see.