Avoid key-UP events on macro selection keys
We want to avoid recording the key-up event of the key pressed to
indicate which macro to record on.  Originally I attempted to avoid this
by masking it in the live_keys array.

This masking just avoids it getting sent to the computer in the HID
events.  It does not avoid our plugin getting given the event.

This patch manually avoids such key-up events in recording (and also
avoids key-up events to choose which key to play from since I noticed it
at the same time).

In implementing this we also happen to avoid recording any key-up event
as the first in a macro.  This only changes any behaviour based on
multiple keys being pressed at the time we choose to replay a macro.
It's odd that the behaviour is not "any key you were holding at the time
does not have its key-up event recorded" but is instead "any key you
were holding at the time does not have its key-up event recorded as the
first action".

However this is something I can look at later.