Handle multiple keys pressed at once
This is something that is supposed to be handled by the MacroSupport
plugin.  Unfortunately I'd missed that the MacroSupport
beforeReportingState hook was not getting called.

Here we follow the lead of the Macros plugin by directly calling it in
our own beforeReportingState hook.

I wonder if this is the best approach, since it would mean that using
multiple macro plugins at the same time would all call this function.
That said, it's done in both DynamicMacros and Macros, and I happen to
not plan to use any other Macro plugin except this one.  Hence there is
no expected problem for the moment.