Added extra tests
N.b. these tests assume that holding a key down during the recording of
a macro keeps that key pressed but *does not* record the key pressing.

I'm about to change that functionality, since I think as a whole it's
too complex and not useful enough to sort out what the behaviour should
be.