Some robustness work.
Make it so that if we run out of macro space when recording the macro is
left empty.
    - Not yet tested.

If there is a key pressed when starting recording, bail.
    - Have made some tests for this.

Special case pressing MACROPLAY when picking slot for recording.
    - This means that MACROREC MACROPLAY will always leave you in
      PICKING_PLAY state (whether you were in the middle of a recording
      or not).  That allows there to be a key sequence that any user
      could always perform in order to ensure the get to the IDLE state.
    - Not yet tested.

Add some folding into the tests file.

Leaving a FAIL in the testsuite since I don't understand the current
behaviour and have run out of time for the day.  The sequence of events
we see seems reasonable, so the current behaviour does not necessarily
need to be changed, but there's certainly something here to figure out.