Small tweak in `MacroSupport::clear`
The current approach means that we send `keyToggledOff` events for a key
*before* removing it from the active_macro_keys_ internal state.

This means that plugins see a keyToggledOff event for a key which *will*
actually be in the next report sent by the HID.

While I don't know of any particular problem that this could cause, it
can be confusing and I expect it could let some plugin writers trip-up.
(I noticed it as strange behaviour in a testcase I was writing for my
local configuration, rather than a bug while using the keyboard).

Here we clear the key in active_macro_keys_ before requesting the
Runtime to handle a release of the key.