Recording "key-pressed and requesting play"
I happened to randomly notice that the code handling
PICKING_SLOT_FOR_PLAY_AND_RECORDING does not have anything to ensure
that holding a key down while replaying avoids the playing.  I made a
random attempt to fix this, but since I've not the time to really test
it this is largely just a reminder.