Use debugging flags and std=c++17 for built testcases
Makes things much easier to debug, and also means I can use c++17
featutres in my testcases.

Not entirely sure whether the C++17 choice is a sensible one given that
I hope to at some point in the future see about getting this upstream
somewhere, but that's a problem for later.