Avoid recursive replay if wasn't run the first time
When recording we avoid playing a macro if there are any keys pressed at
the time that the play is requested.  When replaying we want to avoid
any recursive play if it was avoided during recording.

Unfortunately, the check to see whether any keys are pressed while
recording is based on the globally accessible `live_keys` structure,
whereas there is no real way to check and see whether any keys are
logically "held" while playing a macro.

We could record something special in our macro buffer to indicate that
the first time round this PLAY macro was not actually triggered.
However that seems quite ugly -- we'd need some odd manipulation of
"what was pressed" in order to store information which is already stored
somewhere else.
(During recording we've already recorded "press PLAY button", we'd
either need to sort of clear it out if there are live keys -- or add
some special marker to say "don't actually play").

That's not *too* bad, the macroStorage buffer is effectively an
instruction sequence anyway, and we already plan to do some
"retroactive" manipulation of the buffer for compactness.
But I just think it looks cleaner if we can query things directly.