Consumed keys and recording
When a user chooses a macro slot we want to trigger this plugin but not
also send that keypress across in the HID events.  There are multiple
slight variations on how to consume that keypress.

When keys are not injected, we can use live_keys.mask() to mark that
this key should not be passed along in future HID events.  However,
*only* calling this in the onKeyEvent hook means that
Runtime_::handleKeyEvent will immediately disable what we've done and
mark the relevant live_keys entry as enabled.

In order to avoid that behaviour we have to either change event.addr so
that the event.addr.isValid() check returns `false` or return
EventHandlerResult::ABORT.  Both of these would interfere with other
plugins, but I think that's fine since we are trying to pretend this key
was not pressed.

Unfortunately these actions do not stop MacroSupport recording the
relevant key in active_macro_keys_.  MacroSupport::press also does not
have any easy way to "return" a value from handleKeyEvent, so there
doesn't seem to be an analogous approach to return
EventHandlerReturn::ABORT.

Three possible approaches were considered:
 1. Set a variable on MacroSupport to mark that we did not want
    Key_<whatever> to be stored in active_macro_keys_ on return from
    handleKeyEvent when called in the MacroSupport::press function.
 2. Record a "release" event in our macro immediately after the press
    event, so that the active_macro_keys_ entry is removed.
 3. Accept the behaviour that if a consumed key is held down and other
    keys are pressed after, that consumed key will appear in the HID
    events produced.

Option 1. was considered to "mostly" work.  The problem we have is that
without a major change in the existing Kaleidoscope infrastructure there
seems to be no way to ensure that the relevant key has been pressed from
MacroSupport::press.
live_keys.mask is updated on every cycle, and skips injected keys.
active_macro_keys_ is updated in MacroSupport functions (i.e. not on
every cycle), and does not skip injected keys.
This means there is the unlikely but possible sequence of events which
goes:
 a) MacroSupport::press presses a key (not via MacrosOnTheFly).
 b) This triggers "some other plugin".
 c) "some other plugin" attempts to do something with MacrosOnTheFly
    which consumes a keystroke.
 d) MacrosOnTheFly marks that key should not be recorded.
 e) Eventually we unwind up the call-stack.
 f) MacroSupport::press does not record the relevant key as pressed,
    even though it is.

Option 2. was also considered to "mostly" work.  The main difference
here is that we would end up triggering a bunch of plugins differently
when recording to when replaying.
While it's never great to introduce another difference between recording
and replaying, there are already similar differences.  E.g.
handleKeyswitchEvent hooks are not triggered when replaying, and all
keys are marked as injected.

While option 3. is not very desirable, it is also not a *terrible* event
to happen.  Such a strange behaviour would generally be preferred to
major differences between the behaviour when recording and replaying.

In the end we went with option 2.  This was a tight choice between
option 2. and option 3. since consistency is generally preferred.
However the drastic difference in the likelyhood of observing strange
behaviour was the main reason we ended up making this choice.
(I.e. it was considered much more likely that a user would hold down
their macro selection key while typing something else, than there would
be a strange interaction between an injected Release event and some
plugin).

-------
This choice could easily be proven wrong in the future, if that is then
I believe we could choose option 3. by simply changing maskKeyAndRet to
do nothing except return `EventHandlerResult::ABORT`.