Test around handling recursive playing of macros
This is interesting because it poses a choice around what we should do
when attempting to play the slot that we are currently recording.
As it stands we play what's recorded so far, reach the request to
play that we've already recorded, then complain and exit because that's
a recursive play.

It seems reasonable to wonder whether it would be better to complain as
soon as we've requested playing the macro which is currently getting
recorded.

It would be easy to make changes so we stop the replay happening while
recording (checking against sRecordingSlot).  In order to stop the
replay happening when replaying we would need do special adjustments to
our macro buffer again.

In our testcase we record the following macros:
 A:  does something around attempting to run B.
 B:  does something around attempting to run A.

No matter what it seems like the behaviour that you see when replaying
will be different than the behaviour you see while recording for at
least one ordering.  This is a fundamental thing because when recording
whichever macro is recorded first would see something different.

In order to see what the difference between these choices is, we can do
some case-analysis.  WLOG Say B recorded last.
We block when recording
  When recording:
    B -> A -> block B
  Then when replaying:
    A -> B -> block A
    B -> A -> block B

We do not block when recording (and zero out relevant macro):
  When recording:
    B -> A -> first half of B
  Then when replaying:
    A -> B -> first half of B
    B -> A -> block B

So the choice of whether to block the macro we are currently recording
or not selects between having replaying of your macro look exactly like
recording or not.

Based on the above analysis it would be best if we "zero out" recording
when there was a problem.  If this is done then I believe we would end
up with the invariant that no recursive macros have been recorded
(because any time a macro being recorded creates a recursive macro we
zero out the macro which is getting recorded and ensure this never
happens).

Here we do just that.