Test keyUp is not recorded, and REC->PLAY->REC bailout
Here we add a test that the keyUp of a key which was chosen for the
recording does not trigger any HID event either while recording or while
replaying.

We also add a test that the REC->PLAY->REC bailout which is supposed to
always get us into the IDLE state does so.

In order to add the second test I did a bit of a hack.  I needed access
to the plugin classes state, but testsuites are compiled without the
arduino's automatic inclusion of every plugin that you need.

Rather than try and manually adjust the path, I directly included the
header which describes the layout of the MacrosOnTheFly plugin and
avoided all other headers when compiling tests with some `ifdef`
hackery.

I included those headers for a reason, so I also had to comment out some
functions in the header.  This was not a problem since those functions
were inline and hence would not change the layout of the class.
Hence the testsuite and the rest of the code actually see two different
class declarations, but since they both would expect the `currentState`
member to be in the same place there is no crash.
I.e. I'd be surprised if this were not UB, but I'm pretty sure I
understand what will happen enough to know it should work.  Plus the
only thing that would break is the testsuite, and if that happens I'll
look into it.