Finish recording compression for basic keypresses
While testing I found that we had a bunch of our integer markers in the
wrong place, this commit fixes it.

This commit also adds a few tests which are not yet complete (because I
carried on working without saving at the recommended place and
extracting the different bits of code seems too tiresome).