Fix macro compression when running after a TapCodeSeq
Until this patch we would not always clear the markers indicating where
a tap sequence could be made and where a tap sequence is being recorded.

The case found in this commit where we did not manage to clear these
markers was when we're pressing a key and the last key was not a key
release.  In this case we can not compress *over* this keyrelease, so we
need to clear the `leadingTapCode` and `leadingTapCodeSeq` markers.