Fixes after tests for MACRODELAY
Saw truncated macro recordings because of MACRODELAY.
Partly because we were adding our delayInterval (which was zero, and
hence looked like MACRO_END) after a MACRODELAY for both a release *and*
a keypress after that release (state was not changed from
SETTING_DELAY_AND_RECORDING to IDLE_AND_RECORDING on the key release).

Partly because we were letting a release of MACRODELAY get into the
standard recording, and that was ending up doing some invalid
compression.

In this patch we also tidy up our macros for returning early if a key is
a "non transition" key.  This was added in order to handle a release key
resetting state (for the above mentioned problem).  For consistency we
make the same release key reset state when picking a slot for replaying
or picking a slot for recording into.