Add tests for `PLAY PLAY`
Also changed the values in `sRecordingSlot` and `sLastPlayedSlot` which
would be in the binary itself.  I don't think this is useful in any way
since I should always call `initialise` and that should reset
everything, but having this value correct for "uninitialised" makes me
feel a little better.

N.b. while adding these tests I noticed that the "last played" state was
being carried over tests, which is a little problematic.

To me that implies that most of the MacrosOnTheFly state is being
carried over tests, so that's something to watch out for (n.b. not
double-checked that this is the case).