Test for when we run out of memory
Turns out that I had an unhygenic macro which was causing problems.
Since I'd not put braces around an argument, the expression I used in
its place was being treated differently to how I'd expect.  In this
particular case that meant that the check for "do we have any memory
left" was always returning true!