First steps towards testing my actual config
Have copied across my config file and managed to parse the existing
keyboard setup so I don't have to maintain lists of the same setup.

(As yet still have to copy across changes from my actual config to the
testsuite, but that will have to be fixed on some later date).