Exclude SPECIALSHIFT from live_keys avoidance
This is a little bit of a hack -- I personally only want to avoid one
particular key from this check for PLAY PLAY, I suspect any reasonable
config would only ever be holding down one key in order to press this,
but there's no particular logic behind it.

N.b. for this particular plugin I have a counter of how many times we've
pressed/released the SPECIALSHIFT key.  The whole point is to tell the
difference between having pressed it twice and released it once or
pressed it once and released it once.  When I allow this key in
live_keys when entering a macro play I need to avoid clearing it when
exiting a macro play because otherwise the counter could be subtracted
below zero.

If I'm avoiding releasing this key, then I should also avoid pressing it
in a macro replay (since when releasing I would not know whether a given
key should be released or should not -- it depends on whether the press
was due to the macro replay or due to being held down on play).
To satisfy this desire we put SpecialShift plugin before MacrosOnTheFly.
SpecialShift returns EVENT_CONSUMED when it's seen its relevant key, and
that means that MacrosOnTheFly will simply not see this key event while
recording.