Rename MacrosOnTheFly files to MacroPirate
MacrosOnTheFly is the name of an existing Kaleidoscope plugin that I
took a lot of "how to write a plugin" from.  It had the same
functionality with a few variations.  It also had some extra features
compared to this.

Since this is a different plugin, with different features and different
design choices I feel it should have a different name.

Choice is MacroPirate because it is a Record and Replay functionality
(_aRR_), and because it copies what it watches and redistributes it
later on.