SpecialShift paranoid keyToggledOff and counter == 0 test
I don't think this could happen in usual handling -- and in fact I've
not seen this happen while using this plugin for a long time -- but
other plugins could send any event and other plugins could mask any
event, so for correctness this should be fixed.

It also should mean we have an invariant for the fuzzing -- though not a
particularly useful one, since there's not much code which could break
it.