Reorder headers around in tests
Move standard headers to before other headers.  On Ubuntu 20.04 it seems
the Kaleidoscope.h header eventually includes the below header which has
a `#define` which interacts badly with a standard header.
  .arduino/user/virtual/cores/arduino/Arduino.h

Here we put the standard headers before the project-specific headers
on the assumption the standard headers would be designed not to
interfere with any user code (i.e. using namespaces or
implementation-specific symbol names).